move register/page.js to /app
